Moved view models to phy.plot.
